Conversion to 32-bit floats in FFT lead to conserved quantity drift; switch to 64-bit floats.

+++ ipi/utils/nmtransform.py
@@ -294,15 +294,15 @@ class nm_fft(object):  # ! TODO add (matrix-version) of the open path transforma
         try:
             import pyfftw
             info("Import of PyFFTW successful", verbosity.medium)
-            self.qdummy = pyfftw.n_byte_align_empty((nbeads, 3 * natoms), 16, 'float32')
-            self.qnmdummy = pyfftw.n_byte_align_empty((nbeads // 2 + 1, 3 * natoms), 16, 'complex64')
+            self.qdummy = pyfftw.n_byte_align_empty((nbeads, 3 * natoms), 16, 'float64')
+            self.qnmdummy = pyfftw.n_byte_align_empty((nbeads // 2 + 1, 3 * natoms), 16, 'complex128')
             self.fft = pyfftw.FFTW(self.qdummy, self.qnmdummy, axes=(0,), direction='FFTW_FORWARD')
             self.ifft = pyfftw.FFTW(self.qnmdummy, self.qdummy, axes=(0,), direction='FFTW_BACKWARD')
         except ImportError:  # Uses standard numpy fft library if nothing better
                             # is available
             info("Import of PyFFTW unsuccessful, using NumPy library instead", verbosity.medium)
-            self.qdummy = np.zeros((nbeads, 3 * natoms), dtype='float32')
-            self.qnmdummy = np.zeros((nbeads // 2 + 1, 3 * natoms), dtype='complex64')
+            self.qdummy = np.zeros((nbeads, 3 * natoms), dtype='float64')
+            self.qnmdummy = np.zeros((nbeads // 2 + 1, 3 * natoms), dtype='complex128')
 
             def dummy_fft(self):
                 self.qnmdummy = np.fft.rfft(self.qdummy, axis=0)
